requirements.txt
pandas
numpy
scikit-learn
xgboost
scipy
selenium
beautifulsoup4
nltk
joblib
matplotlib
openai
python-dotenv
sentence-transformers
requests
llama-cpp-python
